+++ packages/freezed/lib/src/templates/concrete_template.dart
@@ -471,17 +471,15 @@ extension DefaultValue on ParameterElement {
 String parseTypeSource(VariableElement element) {
   var type = element.type?.getDisplayString();
 
-  if (type == null || type.contains('dynamic')) {
+  if ((type == null || type.contains('dynamic')) && element.nameOffset > 0) {
+    final source =
+        element.source.contents.data.substring(0, element.nameOffset);
     if (element.type?.element != null &&
         element.type.isDynamic &&
         element.type.element.isSynthetic) {
-      final source =
-          element.source.contents.data.substring(0, element.nameOffset);
       final match = RegExp(r'(\w+)\s+$').firstMatch(source);
       type = match?.group(1);
     } else if (element.type?.element != null) {
-      final source =
-          element.source.contents.data.substring(0, element.nameOffset);
       final match = RegExp(r'(\w+<.+?>)\s+$').firstMatch(source);
       type = match?.group(1) ?? type;
     }