+++ packages/freezed/test/json_test.dart
@@ -37,11 +37,22 @@ Future<void> main() async {
     );
   });
 
-  test('generic json', () {
-    expect(
-      Generic<int>.fromJson(<String, dynamic>{'a': 42}),
-      Generic(42),
-    );
+  group('generic json', () {
+    test('fromJson', () {
+      expect(
+        Generic<int>.fromJson(<String, dynamic>{'a': 42}),
+        Generic(42),
+      );
+    });
+
+    test('tear-off', () {
+      Generic<int> Function(Map<String, Object>) fromJson = $Generic.fromJson;
+
+      expect(
+        fromJson(<String, dynamic>{'a': 42}),
+        Generic(42),
+      );
+    });
   });
 
   test('single ctor + json can access properties/copyWith', () {