+++ packages/freezed/test/multiple_constructors_test.dart
@@ -425,19 +425,31 @@ void main() {
     test('generates List of correct type', () async {
       final nestedListClass = _getClassElement('ShallowNestedList');
 
-      expect(nestedListClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedListClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'List<LeafNestedListItem>');
     });
 
     test('generates List of correct type for deeply nested case', () async {
       final nestedListClass = _getClassElement('DeepNestedList');
 
-      expect(nestedListClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedListClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'List<InnerNestedListItem>');
 
       final nestedListItemClass = _getClassElement('InnerNestedListItem');
 
-      expect(nestedListItemClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedListItemClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'List<LeafNestedListItem>');
     });
   });
@@ -445,19 +457,31 @@ void main() {
     test('generates Map of correct type', () async {
       final nestedMapClass = _getClassElement('ShallowNestedMap');
 
-      expect(nestedMapClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedMapClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'Map<String, LeafNestedMapItem>');
     });
 
     test('generates Map of correct type for deeply nested case', () async {
       final nestedMapClass = _getClassElement('DeepNestedMap');
 
-      expect(nestedMapClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedMapClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'Map<String, InnerNestedMapItem>');
 
       final nestedMapItemClass = _getClassElement('InnerNestedMapItem');
 
-      expect(nestedMapItemClass.getField('children').type.getDisplayString(),
+      expect(
+          nestedMapItemClass
+              .getField('children')
+              .type
+              .getDisplayString(withNullability: false),
           'Map<String, LeafNestedMapItem>');
     });
   });